renamed testresults package - since now it does more than results.
